Logging improvements
Instances of classes that inherit from the new Loggable class can now be logged directly to a Log::Stream. This makes it easier to log objects that provide context for a message.

Additionally, the kqchan code has been updated to use this new functionality to log kqchannels along with messages and make it easier to identify them and their processes.

Finally, for every message, the thread and process logging it is now recorded along with the message.