kqchan: Check for events after sending messages
XNU's kevent filter code returns FILTER_ACTIVE when there are events still available after a filter call. This tells kevent to check for more events at its discretion. Thus, to mimic this behavior, whenever we send a message, we check for more events afterwards.

Note that we only do this once the message has been fully sent. For reasons that are still unclear, checking immediately after queueing the message and queueing a new notification message as a result of this check messes up the order of the messages that our peer sees.