calls: Add console_open
This RPC call gives the caller a socket to which it can write to to log
to the server's log stream.

This is used to give userspace a place to put messages for "/dev/console".
Most notably, launchd tries to log to this device for important log
messages. This allows us to capture those messages.